Use props to display data
By doing that you make components reusable.

Two steps are required:

 - Pass props to components when rendering them;

 - Destructure the `props` objects to access the received props
   individually. You do that in the compononent definition, between the
   parentheses;

Note that "In React, data flows down the component tree. This is
referred to as _one way data flow._ State, which will be discussed in
the next chapter, can be passed from parent to child components as
props.

This note indicates that we can change a parent's data from a child.

You use curly braces to weave in and out of JavaScript and JSX.

Finally, React needs a key for each item because it needs to uniquely
identify them. By doing so, instead of updating the entire list, it can
update individual items.